requirements.txt
opencv-python
ultralytics
pandas
pillow
torch
tk
ttkbootstrap
